Match the type of `n' to source data type `size_t'.

+++ Source/Project64/N64 System/Mips/TLB class.cpp
@@ -303,7 +303,7 @@ void CTLB::RecordDifference( CLog &LogFile, const CTLB& rTLB)
 
 bool CTLB::operator == (const CTLB& rTLB) const
 {
-    for (int i = 0, n = sizeof(m_tlb)/sizeof(m_tlb[0]); i < n; i++)
+    for (size_t i = 0, n = sizeof(m_tlb)/sizeof(m_tlb[0]); i < n; i++)
     {
         if (m_tlb[i].EntryDefined != rTLB.m_tlb[i].EntryDefined)
         {
